ADD: find record number module
script for downloading the information from RIPE stat.
In this script we only consider the number of "events"
which refer to the number of BGP update message during
the select time.

This module is very similar to the as_path_length.

Using the data from RIPE stat is very simple but the catch is
that they don't have all the data available. So make sure to check
first that your query data is indeed exist.